fix audioflinger integer mulplication overflow
when underrun duration is over 3 sec, overflow happens
for multiplication of nsec=sec*1000000000. Use correct format transform
to solve it.

Bug: 135075416
Test: run MTBF test

+++ services/audioflinger/FastThread.cpp
@@ -309,7 +309,7 @@ bool FastThread::threadLoop()
                     // compute the delta value of clock_gettime(CLOCK_MONOTONIC)
                     uint32_t monotonicNs = nsec;
                     if (sec > 0 && sec < 4) {
-                        monotonicNs += sec * 1000000000;
+                        monotonicNs += (uint32_t)sec * (uint32_t)1000000000;
                     }
                     // compute raw CPU load = delta value of clock_gettime(CLOCK_THREAD_CPUTIME_ID)
                     uint32_t loadNs = 0;
@@ -325,7 +325,7 @@ bool FastThread::threadLoop()
                             }
                             loadNs = nsec;
                             if (sec > 0 && sec < 4) {
-                                loadNs += sec * 1000000000;
+                                loadNs += (uint32_t)sec * (uint32_t)1000000000;
                             }
                         } else {
                             // first time through the loop
